fix: clear ci dictation checks
