+++ mobile/lib/src/features/ai_dictation/application/mobile_ai_dictation_controller.dart
@@ -22,6 +22,7 @@ import 'package:speech_to_text/speech_recognition_result.dart';
 import 'package:speech_to_text/speech_to_text.dart';
 
 part 'mobile_ai_dictation_controller.g.dart';
+part 'mobile_ai_dictation_controller_recording.dart';
 part 'mobile_ai_dictation_controller_transcription.dart';
 
 const _capabilityChannel = MethodChannel(
@@ -149,34 +150,6 @@ class MobileAiDictationController extends _$MobileAiDictationController {
       }
     }
     await _deleteRecording();
-  }
-
-  AudioPlayer _ensurePlayer() {
-    final existing = _player;
-    if (existing != null) return existing;
-    final player = AudioPlayer();
-    _player = player;
-    _subscriptions.add(
-      player.positionStream.listen((position) {
-        if (state.stage == MobileAiDictationStage.playing ||
-            state.hasRecording) {
-          state = state.copyWith(playbackPosition: position);
-        }
-      }),
-    );
-    _subscriptions.add(
-      player.playerStateStream.listen((playerState) {
-        if (playerState.processingState == ProcessingState.completed &&
-            state.stage == MobileAiDictationStage.playing) {
-          unawaited(player.seek(Duration.zero));
-          state = state.copyWith(
-            stage: MobileAiDictationStage.recorded,
-            playbackPosition: Duration.zero,
-          );
-        }
-      }),
-    );
-    return player;
   }
 
   Future<void> start({
@@ -347,42 +320,6 @@ class MobileAiDictationController extends _$MobileAiDictationController {
     ).copyWith(segmentCount: _audioPaths.length);
   }
 
-  Future<void> _rotateRecordingSegment(int generation) async {
-    if (!_isCurrentGeneration(generation) ||
-        state.stage != MobileAiDictationStage.recording ||
-        _audioPath == null) {
-      return;
-    }
-    final completedPath = await _recorder?.stop();
-    if (completedPath != null && completedPath != _audioPath) {
-      final index = _audioPaths.indexOf(_audioPath!);
-      if (index >= 0) _audioPaths[index] = completedPath;
-      _audioPath = completedPath;
-    }
-    if (!_isCurrentGeneration(generation) ||
-        state.stage != MobileAiDictationStage.recording) {
-      return;
-    }
-    final directory = await getTemporaryDirectory();
-    final path = p.join(
-      directory.path,
-      'alera-mobile-dictation-${DateTime.now().microsecondsSinceEpoch}-${_audioPaths.length}.wav',
-    );
-    await _recorder?.start(
-      const RecordConfig(
-        encoder: AudioEncoder.wav,
-        sampleRate: 16000,
-        numChannels: 1,
-        autoGain: true,
-        noiseSuppress: true,
-      ),
-      path: path,
-    );
-    _audioPath = path;
-    _audioPaths.add(path);
-    state = state.copyWith(segmentCount: _audioPaths.length);
-  }
-
   void _beginElapsedTimer(int generation) {
     _recordingStarted = DateTime.now();
     _elapsedTimer?.cancel();
@@ -532,16 +469,6 @@ class MobileAiDictationController extends _$MobileAiDictationController {
     _workspaceId = null;
     _tabId = null;
     state = MobileAiDictationState(warning: warning);
-  }
-
-  Future<void> _deleteRecording() async {
-    final paths = _audioPaths.toSet().toList();
-    _audioPaths.clear();
-    _audioPath = null;
-    for (final path in paths) {
-      final file = File(path);
-      if (await file.exists()) await file.delete();
-    }
   }
 }
